feat: make prettier a peer dependency
BREAKING CHANGE
Prettier is now a peer dependency so consumers can choose any version of prettier they'd like.
When there is a new version of prettier out, we don't have to upgrade the configs just to get the upgraded library.
This follows a similar pattern to our eslint/tslint config.